# docs/swagger.yaml
openapi: 3.0.0
info:
  title: ReactViteProKit API
  description: >
    Full-featured backend API built with Node.js, Express, MongoDB, Cloudinary.  
    Modular MVC design with Auth, User, Admin, Product, File, OTP modules.
    All responses follow a standard format with success status, data, and message fields.
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:5000/api
    description: 
  - url: https://example.com/api
    description: 
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token required for authenticated endpoints
paths:
...
  /users/me:
    get:
      tags:
        - User
      summary: Get current user profile
      description: Retrieves the authenticated user's profile information
      operationId: getMyProfile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      _id:
                        type: string
                        description: User unique identifier
                        example: "507f1f77bcf86cd799439011"
                      email:
                        type: string
                        format: email
                        description: User email address
                        example: "user@example.com"
                      phone:
                        type: string
                        description: User phone number
                        example: "+989123456789"
                      role:
                        type: string
                        enum: [user, admin]
                        description: User role in the system
                        example: "user"
                      createdAt:
                        type: string
                        format: date-time
                      updatedAt:
                        type: string
                        format: date-time
                  message:
                    type: string
                    example: "Operation successful"
        "401":
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"

    put:
      tags:
        - User
      summary: Update current user profile
      description: Updates the authenticated user's email, phone, or password
      operationId: editMyProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: New email address
                  example: "newemail@example.com"
                phone:
                  type: string
                  description: New phone number
                  example: "+989123456789"
                password:
                  type: string
                  format: password
                  description: New password
                  example: "NewPassword123!"
      responses:
        "200":
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      _id:
                        type: string
                        description: User unique identifier
                        example: "507f1f77bcf86cd799439011"
                      email:
                        type: string
                        format: email
                        description: User email address
                        example: "user@example.com"
                      phone:
                        type: string
                        description: User phone number
                        example: "+989123456789"
                      role:
                        type: string
                        enum: [user, admin]
                        description: User role in the system
                        example: "user"
                      createdAt:
                        type: string
                        format: date-time
                      updatedAt:
                        type: string
                        format: date-time
                  message:
                    type: string
                    example: "Operation successful"
        "401":
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"
  /users:
    get:
      tags:
        - User
      summary: List all users (admin only)
      description: >
        Retrieves a list of all users in the system.
        Admin only endpoint.
      operationId: getAllUsers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                          description: User unique identifier
                          example: "507f1f77bcf86cd799439011"
                        email:
                          type: string
                          format: email
                          description: User email address
                          example: "user@example.com"
                        phone:
                          type: string
                          description: User phone number
                          example: "+989123456789"
                        role:
                          type: string
                          enum: [user, admin]
                          description: User role in the system
                          example: "user"
                        createdAt:
                          type: string
                          format: date-time
                        updatedAt:
                          type: string
                          format: date-time
                  message:
                    type: string
                    example: "Operation successful"
        "401":
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"
        "403":
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"
  /users/{id}:
    get:
      tags:
        - User
      summary: Get user by ID (admin only)
      description: >
        Retrieves a specific user's profile by ID.
        Admin only endpoint.
      operationId: getUserById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: User ID (MongoDB ObjectId)
          example: "507f1f77bcf86cd799439011"
      responses:
        "200":
          description: User profile retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      _id:
                        type: string
                        description: User unique identifier
                        example: "507f1f77bcf86cd799439011"
                      email:
                        type: string
                        format: email
                        description: User email address
                        example: "user@example.com"
                      phone:
                        type: string
                        description: User phone number
                        example: "+989123456789"
                      role:
                        type: string
                        enum: [user, admin]
                        description: User role in the system
                        example: "user"
                      createdAt:
                        type: string
                        format: date-time
                      updatedAt:
                        type: string
                        format: date-time
                  message:
                    type: string
                    example: "Operation successful"
        "401":
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"
        "403":
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"

    put:
      tags:
        - User
      summary: Update user by ID (admin only)
      description: >
        Updates a specific user's information by ID.
        Admin only endpoint.
      operationId: editUserByAdmin
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: User ID (MongoDB ObjectId)
          example: "507f1f77bcf86cd799439011"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: "newemail@example.com"
                phone:
                  type: string
                  example: "+989123456789"
                password:
                  type: string
                  format: password
                  example: "NewPassword123!"
                role:
                  type: string
                  enum: [user, admin]
                  example: "admin"
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      _id:
                        type: string
                        description: User unique identifier
                        example: "507f1f77bcf86cd799439011"
                      email:
                        type: string
                        format: email
                        description: User email address
                        example: "user@example.com"
                      phone:
                        type: string
                        description: User phone number
                        example: "+989123456789"
                      role:
                        type: string
                        enum: [user, admin]
                        description: User role in the system
                        example: "user"
                      createdAt:
                        type: string
                        format: date-time
                      updatedAt:
                        type: string
                        format: date-time
                  message:
                    type: string
                    example: "Operation successful"
        "401":
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"
        "403":
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"

    delete:
      tags:
        - User
      summary: Delete user by ID (admin only)
      description: >
        Deletes a specific user from the system.
        Admin only endpoint.
      operationId: deleteUserByAdmin
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: User ID (MongoDB ObjectId)
          example: "507f1f77bcf86cd799439011"
      responses:
        "200":
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: "null"
                  message:
                    type: string
                    example: "Operation successful"
        "401":
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"
        "403":
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"
...